Refactor the Chat Message into its own file
